requirements.txt
asgiref==3.2.3
awscli==1.18.5
botocore==1.15.5
colorama==0.4.3
Django==3.0.3
django-crispy-forms==1.9.0
django-multiselectfield==0.1.12
django-richtextfield==1.5.0
django-tinymce==2.8.0
docutils==0.15.2
jmespath==0.9.4
psycopg2==2.8.5
pyasn1==0.4.8
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.2
rsa==3.4.2
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
